Enable docker-in-docker in the CI runners
The app-template CICD pipeline uses appctl to validate and prepare the
manifests. `appctl` depends on `kpt`, which requires a docker daemon to
run in the appctl container.
To solve the docker-in-docker issue, the Kubernetes executor needs to be
in `privileged` mode. Moreover, the stage needs to have some extra
setups.

References:
- https://docs.gitlab.com/runner/executors/kubernetes.html#using-dockerdind
- https://gitlab.com/gitlab-org/gitlab-runner/-/blob/master/common/config.go#L212